refactor: refactored action bar to include action bar base for reusability
