+++ packages/action-bar/exports.json
@@ -1,6 +1,5 @@
 {
     "./src/*": "./src/*",
-    "./scopedActionBar/*": "./scopedActionBar/*",
     "./sp-action-bar.js": "./sp-action-bar.js",
     "./sp-action-bar-scoped.js": "./sp-action-bar-scoped.js"
 }
