+++ packages/action-bar/src/ActionBar.ts
@@ -10,114 +10,13 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import {
-    CSSResultArray,
-    html,
-    SpectrumElement,
-    TemplateResult,
-} from '@spectrum-web-components/base';
-import { property } from '@spectrum-web-components/base/src/decorators.js';
 import '@spectrum-web-components/popover/sp-popover.js';
 import '@spectrum-web-components/action-group/sp-action-group.js';
 import '@spectrum-web-components/button/sp-close-button.js';
 import '@spectrum-web-components/field-label/sp-field-label.js';
-import actionBarStyles from './action-bar.css.js';
-import { ifDefined } from '@spectrum-web-components/base/src/directives.js';
-export const actionBarVariants = ['sticky', 'fixed'];
+import { ActionBarBase } from './ActionBarBase';
 
 /**
  * @element sp-action-bar
- * @slot - Content to display with the Action Bar
  */
-export class ActionBar extends SpectrumElement {
-    public static override get styles(): CSSResultArray {
-        return [actionBarStyles];
-    }
-
-    /**
-     * Deliver the Action Bar with additional visual emphasis.
-     */
-    @property({ type: Boolean, reflect: true })
-    public emphasized = false;
-
-    /**
-     * When `flexible` the action bar sizes itself to its content
-     * rather than a specific width.
-     *
-     * @param {Boolean} flexible
-     */
-    @property({ type: Boolean, reflect: true })
-    public flexible = false;
-
-    @property({ type: Boolean, reflect: true })
-    public open = false;
-
-    /**
-     * The variant applies specific styling when set to `sticky` or `fixed`.
-     * `variant` attribute is removed when not matching one of the above.
-     *
-     * @param {String} variant
-     */
-    @property({ type: String, reflect: true })
-    public set variant(variant: string) {
-        if (variant === this.variant) {
-            return;
-        }
-        if (actionBarVariants.includes(variant)) {
-            this.setAttribute('variant', variant);
-            this._variant = variant;
-            return;
-        }
-        this.removeAttribute('variant');
-        this._variant = '';
-    }
-
-    public get variant(): string {
-        return this._variant;
-    }
-
-    private _variant = '';
-
-    private handleClick(): void {
-        this.open = false;
-
-        const applyDefault = this.dispatchEvent(
-            new Event('close', {
-                bubbles: true,
-            })
-        );
-
-        if (!applyDefault) {
-            this.open = true;
-        }
-    }
-
-    public override render(): TemplateResult {
-        return html`
-            <sp-popover ?open=${this.open} id="popover">
-                <slot name="override">
-                    <sp-close-button
-                        static=${ifDefined(
-                            this.emphasized ? 'white' : undefined
-                        )}
-                        class="close-button"
-                        label="Clear selection"
-                        @click=${this.handleClick}
-                    ></sp-close-button>
-                    <sp-field-label class="field-label">
-                        <slot></slot>
-                    </sp-field-label>
-                    <sp-action-group
-                        class="action-group"
-                        quiet
-                        static=${ifDefined(
-                            this.emphasized ? 'white' : undefined
-                        )}
-                    >
-                        <slot name="buttons"></slot>
-                    </sp-action-group>
-                </slot>
-            </sp-popover>
-        `;
-    }
-}
+export class ActionBar extends ActionBarBase {}
