+++ packages/action-bar/src/ActionBarScoped.ts
@@ -8,7 +8,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { ActionBar as ActionBarBase } from '../src/ActionBarBase';
+import { ActionBarBase } from './ActionBarBase';
 import { ScopedRegistryHost } from '@lit-labs/scoped-registry-mixin';
 import { ActionGroup } from '@spectrum-web-components/action-group';
 import { CloseButton } from '@spectrum-web-components/button';
